update - 스케듈러 등록 및 파일 데이터 저장

+++ src/main/java/com/son/daou/domain/shop/ShopHistory.java
@@ -34,13 +34,13 @@ public class ShopHistory implements DaouEntity {
     Integer deleteCount;
 
     @Column(name = "payment")
-    Integer payment;
+    Long payment;
 
     @Column(name = "used")
-    Integer used;
+    Long used;
 
     @Column(name = "sales")
-    Integer sales;
+    Long sales;
 
     @CreationTimestamp
     @Column(name = "created_at")
